hickory_dns

Struct Config

pub struct Config {
    pub user: Option<String>,
    pub group: Option<String>,
    /* private fields */
}

Server configuration

Fields
user: Option<String>

User to run the server as.

Only supported on Unix-like platforms. If the real or effective UID of the hickory process is root, we will attempt to change to this user (or to nobody if no user is specified here.)

group: Option<String>

Group to run the server as.

Only supported on Unix-like platforms. If the real or effective UID of the hickory process is root, we will attempt to change to this group (or to nobody if no group is specified here.)

Implementations

impl Config

pub fn read_config(path: &Path) -> Result<Self, ConfigError>

read a Config file from the file specified at path.

pub fn from_toml(toml: &str) -> Result<Self, ConfigError>

Read a Config from the given TOML string.

pub fn listen_addrs_ipv4(&self) -> Result<Vec<Ipv4Addr>, AddrParseError>

set of listening ipv4 addresses (for TCP and UDP)

pub fn listen_addrs_ipv6(&self) -> Result<Vec<Ipv6Addr>, AddrParseError>

set of listening ipv6 addresses (for TCP and UDP)

pub fn listen_port(&self) -> u16

port on which to listen for connections on specified addresses

pub fn tls_listen_port(&self) -> u16

port on which to listen for TLS connections

pub fn https_listen_port(&self) -> u16

port on which to listen for HTTPS connections

pub fn quic_listen_port(&self) -> u16

port on which to listen for QUIC connections

pub fn h3_listen_port(&self) -> u16

port on which to listen for HTTP/3 connections

pub fn disable_tcp(&self) -> bool

get if TCP protocol should be disabled

pub fn disable_udp(&self) -> bool

get if UDP protocol should be disabled

pub fn disable_tls(&self) -> bool

get if TLS protocol should be disabled

pub fn disable_https(&self) -> bool

get if HTTPS protocol should be disabled

pub fn disable_quic(&self) -> bool

get if QUIC protocol should be disabled

pub fn tcp_request_timeout(&self) -> Duration

default timeout for all TCP connections before forcibly shutdown

pub fn log_level(&self) -> Level

specify the log level which should be used, [“Trace”, “Debug”, “Info”, “Warn”, “Error”]

pub fn directory(&self) -> &Path

the path for all zone configurations, defaults to /var/named

pub fn zones(&self) -> &[ZoneConfig]

the set of zones which should be loaded

pub fn tls_cert(&self) -> Option<&TlsCertConfig>

the tls certificate to use for accepting tls connections

pub fn deny_networks(&self) -> &[IpNet]

get the networks denied access to this server

pub fn allow_networks(&self) -> &[IpNet]

get the networks allowed to connect to this server