hickory_dns

Struct ServerZoneConfig

pub struct ServerZoneConfig {
    pub allow_axfr: Option<bool>,
    pub stores: Vec<ServerStoreConfig>,
}

Fields
allow_axfr: Option<bool>

Allow AXFR (TODO: need auth)

stores: Vec<ServerStoreConfig>

Store configurations. Note: we specify a default handler to get a Vec containing a StoreConfig::Default, which is used for authoritative file-based zones and legacy sqlite configurations. #[serde(default)] cannot be used, because it will invoke Default for Vec, i.e., an empty Vec and we cannot implement Default for StoreConfig and return a Vec. The custom visitor is used to handle map (single store) or sequence (chained store) configurations.

Implementations

impl ServerZoneConfig

pub fn file(&self) -> Option<&Path>

path to the zone file, i.e. the base set of original records in the zone

this is only used on first load, if dynamic update is enabled for the zone, then the journal file is the actual source of truth for the zone.

pub fn is_axfr_allowed(&self) -> bool

enable AXFR transfers

pub fn is_dnssec_enabled(&self) -> bool

declare that this zone should be signed, see keys for configuration of the keys for signing